Stand In Assets and Shader Added
Imported 2 content packs that contains the shader we are going to work with and as well as good stand in assets to work with in the meantime if we can't get to everything. Placed in a folder called Temp. Also needed to turn on mesh distance feilds in order for the water shader to work properly.